requirements.txt
requests==2.31.0
pandas==2.1.0
google-cloud-bigquery==3.11.4
APScheduler==3.10.4
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-CORS==4.0.0 